refactor(items): store item price as decimal (US201)
- Use Numeric(10,2) for Item.price in model and migration
- Update item schemas to Decimal with precision constraints

+++ alembic/versions/9c0ffd08f892_add_items_and_categories_tables_with_many_to_many_relationship.py
@@ -50,7 +50,7 @@ def upgrade() -> None:
         sa.Column('created_at', sa.DateTime(), nullable=False),
         sa.Column('is_active', sa.Boolean(), nullable=False, server_default=sa.true()),
         sa.Column('description', sa.String(length=500), nullable=True),
-        sa.Column('price', sa.Float(), nullable=True),
+        sa.Column('price', sa.Numeric(10, 2), nullable=True),
         sa.Column('brand', sa.String(length=100), nullable=True),
         sa.Column('image_url', sa.String(length=255), nullable=True),
         sa.Column('company_id', sa.Integer(), nullable=False),
